requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.0
authlib>=1.3.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
